+++ plugins/12_Midnight/localization/frFR.lua
@@ -51,6 +51,7 @@ L['on_flying_platform'] = 'Sur une plate-forme volante.'
 L['gift_of_the_phoenix_note'] = 'Attrapez 5x Cendres de phénix tombant au sol après avoir interagi avec le {object:Réceptacle de Haut-Soleil}.'
 L['triple_locked_safebox_note'] = 'Requiert 3x clés de coffre-fort cachées. Prenez la {object:Torche ardente} à côté du coffre pour les faire apparaître.'
 L['incomplete_book_of_sonnets'] = 'Livret de sonnets incomplet'
+L['ensorcelled_cryptid_note'] = nil
 
 -------------------------------------------------------------------------------
 ----------------------------------- HARANDAR ----------------------------------
@@ -215,4 +216,4 @@ L['options_icons_the_honored_dead_desc'] = 'Afficher les emplacements pour le ha
 L['options_icons_mysterious_mixing'] = '{achievement:63432}'
 L['options_icons_mysterious_mixing_desc'] = 'Afficher les emplacements pour le haut-fait {achievement:63432}.'
 L['mysterious_mixing_note'] = 'Terminez la quête journalière {quest:97016} donnée par {npc:263327} afin de pouvoir fabriquer des offrandes :\n\n{note:Les ingrédients sont obtenables dans les {object:Jarres canope fissurées}, {object:Babioles en venin coagulé}, et {object:Carapaces chantantes} que l\'on trouve sur l\'île, ou achetés directement à l\'hôtel des ventes.}'
-L['curse_surge_note'] = 'Les événements de Vague de malédiction alternent toutes les 45 minutes.\n\n{note:Il est possible de rejoindre un groupe inter-royaumes si vous avez manqué l\'événement, il sera peut-être encore en cours sur cette phase-là.}'
+L['curse_surge_note'] = 'Les événements de Vague de malédiction alternent toutes les %d minutes.\n\n{note:Il est possible de rejoindre un groupe inter-royaumes si vous avez manqué l\'événement, il sera peut-être encore en cours sur cette phase-là.}'